Skip venv setup if already done, add git pull

+++ device/setup.sh
@@ -2,15 +2,20 @@
 
 set -e
 
+git stash
+git pull || echo "git pull failed"
+
 # Add necessary system packages
 sudo DEBIAN_FRONTEND=noninteractive apt-get -y install \
   python3-venv \
   libportaudio0 libportaudio2 libportaudiocpp0 portaudio19-dev \
   libatlas-base-dev
 
-# Set up virtual environment
-rm -rf .venv
-python3 -m venv .venv
+# Set up virtual environment if needed
+if [ ! -d "./venv" ]; then
+  rm -rf .venv
+  python3 -m venv .venv
+fi
 
 # Install Python packages
 . .venv/bin/activate
